requirements.txt
asgiref==3.9.1
Django==4.2.23
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
# Excel export
openpyxl==3.1.2
# PDF export
reportlab==4.0.4
xhtml2pdf==0.2.11
